Added spring osgi tests
